requirements.txt
discord>=2.3.2
discord.py>=2.5.2
google-generativeai>=0.8.4
python-dotenv>=0.21.0
flask>=2.3.3
scikit-learn>=1.6.0
googletrans